:recycle: issue #2 DDDアーキテクチャを意識しながらソースコードをプレゼンテーション・アプリケーション、ドメイン、インフラに分別 - インフラは DB をまだ導入していないので空
