+++ 1weekPDCA/domain/CardView_model.swift
