--- 1weekPDCA/domain/PlanDo/Task_Model.swift
@@ -1,15 +1,0 @@
-//
-//  Task_Model.swift
-//  1weekPDCA
-//
-//  Created by 齊藤虎太郎 on 2023/02/19.
-//
-
-import Foundation
-import SwiftUI
-
-struct Task {
-    let taskName: String
-    let currentProgress: Double
-    let totalProgress: Double
-}
