--- ContentView.swift
@@ -1,27 +1,0 @@
-//
-//  ContentView.swift
-//  1weekPDCA
-//
-//  Created by 齊藤虎太郎 on 2023/01/30.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundColor(.accentColor)
-            Text("Hello, world!")
-                .foregroundColor(Color.uiColorGreen)
-        }
-        .padding()
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
